+++ Sources/Model/digitalproduct.cpp
@@ -1,6 +1,6 @@
 #include "Sources/Headers/digitalproduct.h"
 
-DigitalProduct::DigitalProduct(string name, string descr, string genre, string country, int year, float cost, int stars, string company)
+DigitalProduct::DigitalProduct(string name, string descr, string genre, string country, int year, int cost, int stars, string company)
     : Product(name, descr, genre, country, year, cost, stars), company(company){}
 
 DigitalProduct::~DigitalProduct(){}
