+++ Sources/Model/physicalproduct.cpp
@@ -1,6 +1,6 @@
 #include "Sources/Headers/physicalproduct.h"
 
-PhysicalProduct::PhysicalProduct(string name, string descr, string genre, string country, int year, float cost, int stars, string author)
+PhysicalProduct::PhysicalProduct(string name, string descr, string genre, string country, int year, int cost, int stars, string author)
     :Product(name, descr, genre, country, year, cost, stars), author(author){}
 
 PhysicalProduct::~PhysicalProduct(){}
